+++ mgplvm/optimisers/__init__.py
@@ -1,0 +1,2 @@
+from .stopping_criterions import (LossMarginStop)
+from . import (sgp, svgp)
