+++ tests/test_kernels.py
@@ -3,7 +3,7 @@ from mgplvm.kernels import QuadExp, QuadExpARD, Linear, Matern
 import numpy as np
 from torch import optim
 import mgplvm
-from mgplvm import rdist, models, training, syndata, likelihoods, lpriors
+from mgplvm import rdist, models, optimisers, syndata, likelihoods, lpriors
 from mgplvm.manifolds import Torus, Euclid
 torch.set_default_dtype(torch.float64)
 
@@ -89,16 +89,16 @@ def test_kernels_run():
                              whiten=True).to(device)
 
         ### test that training runs ###
-        trained_mod = training.svgp(Y,
-                                    mod,
-                                    device,
-                                    optimizer=optim.Adam,
-                                    max_steps=5,
-                                    burnin=100,
-                                    n_mc=64,
-                                    lrate=10E-2,
-                                    print_every=50,
-                                    n_svgp=0)
+        trained_mod = optimisers.svgp.optimise(Y,
+                                               mod,
+                                               device,
+                                               optimizer=optim.Adam,
+                                               max_steps=5,
+                                               burnin=100,
+                                               n_mc=64,
+                                               lrate=10E-2,
+                                               print_every=50,
+                                               n_svgp=0)
 
     return
